WIP Make 'react-dom-factories' publishable
**what is the change?:**
First steps towards making this package something we can publish on 'npm' and use in a variety of contexts.

**why make this change?:**
We need to get this published to unblock the 15.6 React release.

**test plan:**
None yet.

**issue:**
https://github.com/facebook/react/issues/9398
