+++ packages/react-dom-factories/package.json
@@ -13,7 +13,7 @@
   },
   "homepage": "https://facebook.github.io/react/",
   "peerDependencies": {
-    "react": "^16.0.0-alpha"
+    "react": "^15.5.4"
   },
   "files": [
     "LICENSE",
